Move Nary and Unary under a new Embed hierarchy
We will be adding `Binary` shortly.
